Fix text repetition issue
When I first wrote this code I did not know how text was formatted on
the display. Now I realize that you have to set the cursor position
before writing or it overwrites what has already been written in an odd
way. This commit fixes writing text to the display so that it does this
correctly.