Fix syntax issue and value scaling
Fix issue where the incorrect define was used for the DataDog IP. Also
add data scaling. The raw sensor value is still published to Datadog,
but the value put on the display is LOW or HIGH if the reported value is
below 500 or above 700, respectively. Otherwise, the value is properly
scaled between those values.